Add checkout-to-workspace command
Using the checkout step with the default executor fails, because git and
ssh are not installed.

The checkout-to-workspace command can be used with the new scm executor
to check out the source once at the start and persist it to the
workspace. This way all sentry-cli commands just need to attach the
workspace to have the repository available.